Issue 1, OTU Filtering
I uploaded 5 documents that detail the otus that were removed by my
filtering process. There is an excel document that has the OTU id and
taxonomy of the otus that were removed. The other four text files are
the otu tables of only negative control otu's with all samples and
negative controls.