Issue 2 - Variation in alpha diversity
There are 3 sub-directories within this directory. The first contains
plots of the coefficeint of variation for 3 alpha diversity metrics.
The x-axis in each are: red=forehead, brown=gut, blue=palm,
green=tongue. Black bars are the means for each body habitat. These
results still need to be statistically tested (ANOVA) but the general
trend that variability of palms > foreheads > guts > tongues still
appears to be true. It should be noted that each point represents one
individual and was calculated using 5 randomly selected time points for
each individual. The second directory contains histograms of the
distributions for all and each individual(s). The third directory
contians plots of alpha diversity through time for each individual. The
red line in each of these plots is the mean for that individual.