Add preliminary work for trajectory analysis
This work covers the statistical analysis of the trajectories based on
the available metadata. The algorithms used are avg, trajectory, diff
and wdiff (with a window size of 2).

The file called processed.tgz contains eps files showing the
trajectories that clustered together per body site. Note that this is
not final and this tries to only prove the type of information that can
be extracted.